requirements.txt
numpy<2
opencv-python
matplotlib
scikit-learn
pillow
flask

tensorflow
